Added option to skip downloading file if file with same name already exists.

+++ YoutubeDownloader/YTMultiDownloader.cs
@@ -16,6 +16,7 @@ namespace YoutubeDownloader
         private string ExportVideoDirPath { get; set; }
         private string ExportAudioDirPath { get; set; }
         private ExportOptions ExportOptions { get; set; }
+        private bool SkipVideosWhichExists { get; set; }
         #endregion
 
         private IDictionary<Guid, Engine> engines;
@@ -66,6 +67,7 @@ namespace YoutubeDownloader
             this.ExportAudioDirPath = builder.ExportAudioDirPath;
             this.ExportVideoDirPath = builder.ExportVideoDirPath;
             this.ExportOptions = builder.ExportOptions;
+            this.SkipVideosWhichExists = builder.SkipVideosWhichExists;
         }
 
         public DownloadResult[] DownloadMulti()
@@ -84,20 +86,25 @@ namespace YoutubeDownloader
             {
                 Task<DownloadResult> downloadTask = Task.Run(() =>
                 {
-                    string vidExt;
                     string videoName;
+                    string videoFilePath;
 
-                    string videoFilePath = DownloadVideo(linkInfo, out videoName, out vidExt);
+                    var videoInfo = GetVideoInfo(linkInfo, out videoName, out videoFilePath);
+
+                    if (File.Exists(videoFilePath) && SkipVideosWhichExists)
+                        return new DownloadResult() { VideoSavedFilePath = videoFilePath, GUID = linkInfo.GUID, AudioSavedFilePath = null, FileBaseName = videoName, DownloadSkipped = true };
+
+                    DownloadVideo(videoInfo, linkInfo.GUID, videoFilePath);
 
                     if (ExportOptions.HasFlag(ExportOptions.ExportAudio))
-                        DownloadAudio(linkInfo, videoFilePath, vidExt);
+                        DownloadAudio(linkInfo, videoFilePath, videoInfo.VideoExtension);
 
                     if (ExportOptions.HasFlag(ExportOptions.ExportAudio) && !ExportOptions.HasFlag(ExportOptions.ExportVideo))
                     {
                         File.Delete(videoFilePath);
                     }
 
-                    return new DownloadResult() { VideoSavedFilePath = videoFilePath, GUID = linkInfo.GUID, AudioSavedFilePath = TransFormToAudioPath(videoFilePath, vidExt), FileBaseName = videoName }; ;
+                    return new DownloadResult() { VideoSavedFilePath = videoFilePath, GUID = linkInfo.GUID, AudioSavedFilePath = TransFormToAudioPath(videoFilePath, videoInfo.VideoExtension), FileBaseName = videoName, DownloadSkipped = false };
                 });
 
                 tasks.Add(downloadTask);
@@ -105,34 +112,35 @@ namespace YoutubeDownloader
             return await Task.WhenAll(tasks);
         }
 
-        private string DownloadVideo(LinkInfo linkInfo, out string videoName, out string videoExtension)
+        private VideoInfo GetVideoInfo(LinkInfo linkInfo, out string videoName, out string videoFilePath)
         {
             IEnumerable<VideoInfo> videoInfos = DownloadUrlResolver.GetDownloadUrls(linkInfo.URL, false);
 
             // Select the first video by type with highest AudioBitrate
-            VideoInfo video = videoInfos
+            VideoInfo videoInfo = videoInfos
                 .OrderByDescending(info => info.AudioBitrate)
                 .First(info => info.VideoType == linkInfo.VideoType);
 
             // This is must, cause we decrypting only this video
-            if (video.RequiresDecryption)
+            if (videoInfo.RequiresDecryption)
             {
-                DownloadUrlResolver.DecryptDownloadUrl(video);
+                DownloadUrlResolver.DecryptDownloadUrl(videoInfo);
             }
 
-            string fileBaseName = linkInfo.FileName == null ? video.Title.ToSafeFileName() : linkInfo.FileName.ToSafeFileName();
-            var filePath = Path.Combine(ExportVideoDirPath, fileBaseName + video.VideoExtension);
-
-            var videoDownloader = videoDownloaders[linkInfo.GUID];
-
-            videoDownloader.Init(video, filePath);
-            videoDownloader.Execute();
+            string fileBaseName = linkInfo.FileName == null ? videoInfo.Title.ToSafeFileName() : linkInfo.FileName.ToSafeFileName();
 
             videoName = fileBaseName;
-            videoExtension = video.VideoExtension;
+            videoFilePath = Path.Combine(ExportVideoDirPath, fileBaseName + videoInfo.VideoExtension);
 
-            // full video path
-            return String.Concat(Path.Combine(ExportVideoDirPath, fileBaseName), video.VideoExtension); ;
+            return videoInfo;
+        }
+
+        private void DownloadVideo(VideoInfo videoInfo, Guid linkGUID, string videoFilePath)
+        {
+            var videoDownloader = videoDownloaders[linkGUID];
+
+            videoDownloader.Init(videoInfo, videoFilePath);
+            videoDownloader.Execute();
         }
 
         private void DownloadAudio(LinkInfo linkInfo, string videoFilePath, string videoExtension)
@@ -152,7 +160,7 @@ namespace YoutubeDownloader
         {
             string audioOutputPath = videoFilePath
                                     .Replace(ExportVideoDirPath, ExportAudioDirPath)
-                                    .ReplaceLastOccurrence(videoExtension, "mp3");
+                                    .ReplaceLastOccurrence(videoExtension, ".mp3");
             return audioOutputPath;
         }
     }
